Implement create, cache flush, fix symbols
The type uses ruby symbols for the name, but the provider was using
strings inside the inspect method.

The create method did not get all of the parameters it needed from the
resource object.

Instances were not being destroyed properly.

Flush was always getting called even if the object wasn't "dirty."